refactor: move ZeroPreserving.hpp to expression/unary/detail/
ZeroPreserving.hpp contains internal traits and concepts, not expression
types. It belongs in the detail/ subdirectory alongside other unary
implementation helpers like CoeffWiseTraits.hpp.
