+++ include/zipper/expression/unary/CoefficientWiseOperation.hpp
@@ -2,7 +2,7 @@
 #define ZIPPER_EXPRESSION_UNARY_COEFFICIENTWISE_OPERATION_HPP
 
 #include "UnaryExpressionBase.hpp"
-#include "ZeroPreserving.hpp"
+#include "detail/ZeroPreserving.hpp"
 #include "zipper/expression/detail/ExpressionTraits.hpp"
 #include "zipper/expression/detail/IndexSet.hpp"
