+++ include/zipper/expression/unary/math/Asin.hpp
@@ -2,7 +2,7 @@
 #define ZIPPER_EXPRESSION_UNARY_MATH_ASIN_HPP
 
 #include "../CoefficientWiseOperation.hpp"
-#include "../ZeroPreserving.hpp"
+#include "../detail/ZeroPreserving.hpp"
 #include <cmath>
 
 namespace zipper::expression::unary {
