+++ include/zipper/expression/unary/math/Round.hpp
@@ -2,7 +2,7 @@
 #define ZIPPER_EXPRESSION_UNARY_MATH_ROUND_HPP
 
 #include "../CoefficientWiseOperation.hpp"
-#include "../ZeroPreserving.hpp"
+#include "../detail/ZeroPreserving.hpp"
 #include <cmath>
 
 namespace zipper::expression::unary {
