+++ include/zipper/expression/unary/math/Tan.hpp
@@ -2,7 +2,7 @@
 #define ZIPPER_EXPRESSION_UNARY_MATH_TAN_HPP
 
 #include "../CoefficientWiseOperation.hpp"
-#include "../ZeroPreserving.hpp"
+#include "../detail/ZeroPreserving.hpp"
 #include <cmath>
 
 namespace zipper::expression::unary {
